+++ R/utils.R
@@ -62,7 +62,7 @@ text_green <- function(text) { paste0('\033[32m',text,'\033[39m') }
 
 #' Calculate the mode of a provided vector
 #'
-#' @param A [`vector`] of values or characters.
+#' @param x A [`vector`] of values or characters.
 #' @param na.rm [`logical`] whether \code{NA} values are to be removed (Default: \code{TRUE})
 #'
 #' @keywords utils, misc